+++ ruoyi-system/src/main/java/com/ruoyi/project/ledger/domain/bo/MeaLedgerBreakdownDetailBo.java
@@ -88,7 +88,7 @@ public class MeaLedgerBreakdownDetailBo extends BaseEntity {
     /**
      * 变更分解数量
      */
-    private Long bgfjsl;
+    private BigDecimal bgfjsl;
     /**
      * 复核数量
      */
